Update files for preprocessing without headings, embeddings, rag

+++ ai/embeddings/Code/centralize-ALL-files.py
@@ -11,5 +11,7 @@ for method_dir in os.listdir(base_dir):
         if filename.endswith(".parquet") and "ALL" in filename:
             path = os.path.join(method_path, filename)
             dest_path = f"../SP800/{(''.join(method_dir.split('-')[:-1])).replace('_', '')}-{filename}"
+            if os.path.exists(dest_path):
+                continue
             shutil.copy2(path, dest_path)
             print(f"Copied file {path} to {dest_path}")
